Add files via upload
This is a front-end web page developed with HTML, CSS, and JavaScript. The project emphasizes clean, responsive design and user-friendly navigation. It serves as an excellent example of how to structure and style a modern web page.

This project is perfect for both learning and showcasing your front-end development skills,